+++ prisma/migrations/20240106131428_added_permissions_to_access/migration.sql
@@ -1,0 +1,5 @@
+-- CreateEnum
+CREATE TYPE "AccessPermission" AS ENUM ('READ', 'READ_RESTRICTED');
+
+-- AlterTable
+ALTER TABLE "Access" ADD COLUMN "permissions" "AccessPermission"[] DEFAULT ARRAY['READ_RESTRICTED']::"AccessPermission"[];