+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `permissions` to the `Access` model
+
 ### Changed
 
 - Improved the language localization for German (`de`)